refactor(web): break up appeal mapper
refactor(web): create wee files for all the wee functions

refactor(web): move has out to its own file

refactor(web): remove split script

feature(web): actually do the thing the ticket was about

refactor(web): ignore types in test, address comment

refactor(web): ignore type errors in timetable submappers for now

refactor(web): wip: groundwork for de-asyncing submappers

refactor(web): wip: groundwork for de-asyncing submappers

refactor(web): de-asyncify submappers

refactor(web): update typedef, remove console and todos

refactor(web): rename s78 -> fpa

refactor(web): fix type error

refactor(web): fix type error

refactor(a2-57): attempt to unwind some conflicting things
