+++ appeals/web/src/server/lib/mappers/appeal/submappers/appeal-reference.mapper.js
@@ -1,0 +1,20 @@
+/**
+ * @typedef {import('#appeals/appeal-details/appeal-details.types.js').WebAppeal} WebAppeal
+ * @typedef {import('../../../../app/auth/auth-session.service.js').SessionWithAuth} SessionWithAuth
+ */
+
+/** @type {import('../appeal.mapper.js').SubMapper} */
+export const mapAppealReference = ({ appealDetails }) => ({
+	id: 'appeal-reference',
+	display: {
+		summaryListItem: {
+			key: {
+				text: 'Appeal reference'
+			},
+			value: {
+				text: appealDetails.appealReference
+			},
+			classes: 'appeal-appeal-reference'
+		}
+	}
+});
