+++ appeals/web/src/server/lib/mappers/appeal/submappers/appeal-type.mapper.js
@@ -1,0 +1,12 @@
+import { textDisplayField } from '#lib/page-components/text.js';
+
+/** @type {import('../appeal.mapper.js').SubMapper} */
+export const mapAppealType = ({ appealDetails, currentRoute, userHasUpdateCasePermission }) =>
+	textDisplayField({
+		id: 'appeal-type',
+		text: 'Appeal type',
+		value: appealDetails.appealType || 'No appeal type',
+		link: `${currentRoute}/change-appeal-type/appeal-type`,
+		userHasEditPermission: userHasUpdateCasePermission,
+		classes: 'appeal-appeal-type'
+	});
