+++ appeals/web/src/server/lib/mappers/appeal/submappers/lead-or-child.mapper.js
@@ -1,0 +1,18 @@
+import { linkedAppealStatus } from '#lib/appeals-formatter.js';
+
+/** @type {import('../appeal.mapper.js').SubMapper} */
+export const mapLeadOrChild = ({ appealDetails }) => ({
+	id: 'lead-or-child',
+	display:
+		appealDetails.linkedAppeals.length > 0
+			? {
+					statusTag: {
+						status: linkedAppealStatus(
+							appealDetails.isParentAppeal || false,
+							appealDetails.isChildAppeal || false
+						),
+						classes: 'govuk-!-margin-left-1 govuk-!-margin-bottom-4'
+					}
+			  }
+			: {}
+});
