+++ appeals/web/src/server/lib/mappers/appeal/submappers/other-appeals.mapper.js
@@ -1,0 +1,48 @@
+import { permissionNames } from '#environment/permissions.js';
+import * as displayPageFormatter from '#lib/display-page-formatter.js';
+import { mapActionComponent } from '../../permissions.mapper.js';
+
+/** @type {import('../appeal.mapper.js').SubMapper} */
+export const mapOtherAppeals = ({ appealDetails, currentRoute, session }) => {
+	const otherAppealsItems = [];
+
+	if (appealDetails.otherAppeals.length) {
+		otherAppealsItems.push(
+			mapActionComponent(permissionNames.updateCase, session, {
+				text: 'Manage',
+				href: `${currentRoute}/other-appeals/manage`,
+				visuallyHiddenText: 'Related appeals',
+				attributes: { 'data-cy': 'manage-related-appeals' }
+			})
+		);
+	}
+
+	otherAppealsItems.push(
+		mapActionComponent(permissionNames.updateCase, session, {
+			text: 'Add',
+			href: `${currentRoute}/other-appeals/add`,
+			visuallyHiddenText: 'Related appeals',
+			attributes: { 'data-cy': 'add-related-appeals' }
+		})
+	);
+
+	return {
+		id: 'other-appeals',
+		display: {
+			summaryListItem: {
+				key: {
+					text: 'Related appeals'
+				},
+				value: {
+					html:
+						displayPageFormatter.formatListOfRelatedAppeals(appealDetails.otherAppeals || []) ||
+						'<span>No related appeals</span>'
+				},
+				actions: {
+					items: otherAppealsItems
+				},
+				classes: 'appeal-other-appeals'
+			}
+		}
+	};
+};
